requirements.txt
fastapi
uvicorn
playwright
beautifulsoup4
lxml
playwright-stealth
pdfplumber
requests
python-dotenv
